% Generated by roxygen2 (4.0.1): do not edit by hand
\name{citrus.calculateFeature.type}
\alias{citrus.calculateFeature.abundances}
\alias{citrus.calculateFeature.medians}
\alias{citrus.calculateFeature.relativeAbundances}
\alias{citrus.calculateFeature.type}
\title{Calculate descriptive cluster features}
\usage{
citrus.calculateFeature.abundances(clusterIds, clusterAssignments,
  citrus.combinedFCSSet, fileIds = NULL, ...)

citrus.calculateFeature.medians(clusterIds, clusterAssignments,
  citrus.combinedFCSSet, medianColumns, fileIds = NULL, ...)

citrus.calculateFeature.relativeAbundances(clusterIds, clusterAssignments,
  citrus.combinedFCSSet, fileIds = NULL, ...)
}
\arguments{
\item{clusterIds}{Cluster IDs that descriptive features should be calculated for.}

\item{clusterAssignments}{List with indicies of cells belonging to each cluster.}

\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object.}

\item{fileIds}{Vector of file IDs to calculate features for. If \code{NULL}, calculates
features for all samples in \code{citrus.combinedFCSSet}.}

\item{medianColumns}{Vector of parameter names or numeric indicies of parameters for which to calculate cluster median values for.}

\item{...}{Other arguments (ignored).}
}
\description{
Calculate descriptive cluster features.
}
\details{
See \link{citrus.calculateFeatures} for examples.
}
\author{
Robert Bruggner
}
\seealso{
\link{citrus.calculateFeatures}, \link{citrus.calculateFoldFeatureSet}
}

