% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.calculateFoldFeatureSet}
\alias{citrus.calculateFoldFeatureSet}
\title{Build cluster features for folds of clustering}
\usage{
citrus.calculateFoldFeatureSet(citrus.foldClustering, citrus.combinedFCSSet,
  featureType = "abundances", minimumClusterSizePercent = 0.05, ...)
}
\arguments{
\item{citrus.foldClustering}{A \code{citrus.foldClustering} object}

\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object}

\item{featureType}{Type of feature to be calculated. Valid options are: \code{abundances} and \code{medians}. See \code{\link{citrus.calculateFeatures}} for additional argument details.}

\item{minimumClusterSizePercent}{Minimum cluster size percent used to select clusters for analysis. See \code{\link{citrus.selectClusters}}.}

\item{...}{Additional arguments passed to feature-type specific calculation functions.}
}
\value{
A \code{citrus.foldFeatureSet} object with properties:
\item{foldLargeEnoughClusters}{List of selected clusters for each fold of clustering.}
\item{foldFeatures}{List of features constructed from fold clustered samples.}
\item{leftoutFeatures}{List of features constructed from non-clustered samples that were mapped to the fold clustering space.}
\item{allLargeEnoughClusters}{Selected clusters from clustering of all samples.}
\item{allFeatures}{Features constructed from clustering of all samples.}
\item{minimumClusterSizePercent}{User-specified minimum cluster size percent.}
\item{folds}{List of sample folds.}
\item{nFolds}{Number of folds.}
}
\description{
Build cluster features for each fold of clustering. If multiple folds of clustering have been performed, \code{citrus.calculateFoldFeatureSet}
builds features for clustered and leftout samples for each fold.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List disease group of each sample
labels = factor(rep(c("Healthy","Diseased"),each=10))

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster each fold
citrus.foldClustering = citrus.clusterAndMapFolds(citrus.combinedFCSSet,clusteringColumns,labels,nFolds=4)

# Build fold features and leftout features
citrus.foldFeatureSet = citrus.calculateFoldFeatureSet(citrus.foldClustering,citrus.combinedFCSSet)
}
\author{
Robert Bruggner
}

