% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.checkFileParameterConsistency}
\alias{citrus.checkFileParameterConsistency}
\title{Checks consistency of FCS File parameters.}
\usage{
citrus.checkFileParameterConsistency(dataDirectory)
}
\arguments{
\item{dataDirectory}{Full path to directory where FCS files are.}
}
\value{
Returns \code{0} if file parameters are consistent, \code{-1} otherwise.
}
\description{
Checks consistency of FCS File parameter names. Will report if FCS files have differing numbers of
parameters or parameters have different names.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Check parameter consistency.
 citrus.checkFileParameterConsistency(dataDirectory)
}
\author{
Robert Bruggner
}

