% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.checkFileParameterConsistencyUI}
\alias{citrus.checkFileParameterConsistencyUI}
\title{Graphical Interface to checking file parameter consistency.}
\usage{
citrus.checkFileParameterConsistencyUI()
}
\description{
Graphical Interface to checking file parameter consistency.
}
\details{
When running \code{citrus.checkFileParameterConsistencyUI()}, the user will first be prompted to select a file.
The user should then select a single FCS file from the directory where all FCS files reside. Citrus will then check the
consistency of parameter names and counts for all FCS files in that directory.
}
\author{
Robert Bruggner
}

