% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.cluster}
\alias{citrus.cluster}
\alias{print.citrus.clustering}
\title{Cluster a \code{citrus.combinedFCSSet}}
\usage{
citrus.cluster(citrus.combinedFCSSet, clusteringColumns,
  clusteringType = "hierarchical", ...)

print.citrus.clustering(citrus.clustering)
}
\arguments{
\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object.}

\item{clusteringColumns}{Vector of names or indicies of data to be used for clustering.}

\item{clusteringType}{Type of clustering to be perfomed. Valid options are: \code{hierarchical}.}

\item{...}{Other arguments passed to specific clustering algorithm.}
}
\value{
A \code{citrus.clustering} object.
\item{clustering}{Clustering object returned by clustering algorithm.}
\item{clusterMembership}{List, containing indicies of cells in \code{citrus.combinedFCSSet} data that belong to each cluster.}
\item{type}{Type of clustering performed.}
\item{clusteringColumns}{Parameters used for clustering}
}
\description{
Cluster data in a \code{citrus.combinedFCS} set object
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster data
citrus.clustering = citrus.cluster(citrus.combinedFCSSet,clusteringColumns)
}
\author{
Robert Bruggner
}

