% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.clusterAndMapFolds}
\alias{citrus.clusterAndMapFolds}
\alias{print.citrus.foldClustering}
\title{Cluster independent folds of data}
\usage{
citrus.clusterAndMapFolds(citrus.combinedFCSSet, clusteringColumns,
  labels = NULL, nFolds = 1, ...)

print.citrus.foldClustering(citrus.foldClustering)
}
\arguments{
\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object.}

\item{clusteringColumns}{Vector of parameter names or indicies to be used for clustering.}

\item{labels}{Labels of samples being clustered. If supplied, used for balancing folds for clustering}

\item{nFolds}{Number of independent folds of clustering to perform. If \code{nFolds=1}, all data are
clustered together and model is regression model is constructed from single feature set.}

\item{...}{Other arguments passed to specific clustering functions.}
}
\value{
A \code{citrus.foldClustering} object
\item{folds}{Indicies of sample rows to be omitted during each fold of clustering. Only defined if \code{nFolds > 1}.}
\item{foldClustering}{\code{citrus.clustering} objects for each fold. Only defined if \code{nFolds > 1}.}
\item{foldMappingAssignments}{\code{citrus.mapping} objects for each fold, containing mapping of data from left-out samples. Only defined if \code{nFolds > 1}}
\item{allClustering}{\code{citrus.clusteringObject} from all sample data.}
\item{nFolds}{Number of independent folds.}
}
\description{
Cluster subsets of data from different samples and maps leftout sample data to fold cluster space.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List disease group of each sample
labels = factor(rep(c("Healthy","Diseased"),each=10))

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster each fold
citrus.foldClustering = citrus.clusterAndMapFolds(citrus.combinedFCSSet,clusteringColumns,labels,nFolds=4)
}
\author{
Robert Bruggner
}

