% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.createHierarchyGraph}
\alias{citrus.createHierarchyGraph}
\title{Create a graph object from clustering hierarchy}
\usage{
citrus.createHierarchyGraph(citrus.clustering, selectedClusters)
}
\arguments{
\item{citrus.foldFeatureSet}{A \code{citrus.foldFeatureSet} object. Clusters for which features are calculated are included in the graph.}

\item{citrus.foldClustering}{A \code{citrus.foldClustering} object. Used to determine relationships between clusters.}
}
\value{
A \code{graph} object and other properties for plotting.
\item{graph}{An \code{\link{igraph}} \code{graph} object.}
\item{layout}{An \code{\link{igraph}} \code{layout} for the graph.}
\item{plotSize}{Size of PDF for plotting (inches) calculated based on the number of clusters to be plotted.}
}
\description{
Create a graph object from clustering hierarchy that may be used by other plotting functions.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster data
citrus.clustering = citrus.cluster(citrus.combinedFCSSet,clusteringColumns)

# Large enough clusters
largeEnoughClusters = citrus.selectClusters(citrus.clustering)

# Create graph for plotting
hierarchyGraphStuff = citrus.createHierarchyGraph(citrus.clustering,selectedClusters=largeEnoughClusters)
}
\author{
Robert Bruggner
}

