% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.exportCluster}
\alias{citrus.exportCluster}
\title{Exports cluster events to file}
\usage{
citrus.exportCluster(clusterId, citrus.clustering, citrus.combinedFCSSet,
  outputDirectory, conditions = NULL)
}
\arguments{
\item{clusterId}{ID of cluster to be exported}

\item{citrus.clustering}{\code{citrus.clustering} object to export clusters from.}

\item{citrus.combinedFCSSet}{\code{citrus.combinedFCSSet} object that contains data to be exported.}

\item{outputDirectory}{Where cluster data should be exported to.}

\item{conditions}{If not \code{NULL}, only export data from the specified conditions.}
}
\description{
Exports cluster events from each file
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster data
citrus.clustering = citrus.cluster(citrus.combinedFCSSet,clusteringColumns)

# Export Cluster
# citrus.exportCluster(clusterId=19500,citrus.clustering,citrus.combinedFCSSset,outputDirectory="/PATH/TO/DIRECTORY/")
}
\author{
Robert Bruggner
}

