% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.exportClusterUI}
\alias{citrus.exportClusterUI}
\title{UI for exporting file cluster events}
\usage{
citrus.exportClusterUI()
}
\description{
UI for exporting file cluster events. When run, a selection dialog will prompt the user to select
a file. The user should select a "citrusClustering.rData" file from a citrusOutput directory.
The user should then enter a comma-separated list of clusters IDs to be exported. A directory
called 'clusterExportData' will be created in the same directory as the selected clustering file
and exported clusters will be placed there.
}
\author{
Robert Bruggner
}

