% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.familyList}
\alias{citrus.familyList}
\title{List possible model families}
\usage{
citrus.familyList()
}
\value{
Vector of model families that Citrus can compute.
}
\description{
Returns valid model family types that Citrus can compute.
}
\author{
Robert Bruggner
}

