% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.featureTypes}
\alias{citrus.featureTypes}
\title{List of computable feature types}
\usage{
citrus.featureTypes()
}
\value{
Vector of feature types Citrus can compute.
}
\description{
Returns valid types of descriptive features that Citrus can compute
}
\author{
Robert Bruggner
}

