% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.fileEventCount}
\alias{citrus.fileEventCount}
\title{Counts FCS File events}
\usage{
citrus.fileEventCount(dataDirectory, ...)
}
\arguments{
\item{dataDirectory}{Directory from which to read FCS files.}

\item{...}{Other parameters passed to citrus.readFCS}
}
\value{
A matrix detailing the number of events and parameters in each directory FCS file.
}
\description{
Counts FCS events in all FCS files in a data directory
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Get FCS Event Counts
citrus.fileEventCount(dataDirectory)
}
\author{
Robet Bruggner
}

