% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.getRelatedClusterIds}
\alias{citrus.getClusterAncestors}
\alias{citrus.getClusterDecendants}
\alias{citrus.getRelatedClusterIds}
\title{Get ancestors or decendants of a cluster}
\usage{
citrus.getClusterDecendants(clusterId, mergeOrder)

citrus.getClusterAncestors(clusterId, mergeOrder)
}
\arguments{
\item{clusterId}{ID of a cluster for which to retreive related clusters.}

\item{mergeOrder}{\code{mergeOrder} result from \code{hclust} object.}
}
\value{
Vector of cluster ids that are decendants of \code{clusterId} argument.
}
\description{
Get ancestors or decendants of a cluster for hierarchical clustering
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster data
citrus.clustering = citrus.cluster(citrus.combinedFCSSet,clusteringColumns)

# Get decendants
citrus.getClusterDecendants(15000,citrus.clustering$clustering$merge)

# Get ancestors
citrus.getClusterAncestors(15000,citrus.clustering$clustering$merge)
}
\author{
Robert Bruggner
}

