% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.launchUI}
\alias{citrus.launchUI}
\title{Launch web-based interface for configuring and running citrus}
\usage{
citrus.launchUI(dataDirectory = NULL)
}
\arguments{
\item{dataDirectory}{If specified, launches configuration UI with files in data directory. If \code{NULL},
prompts user to select a single FCS file in data directory.}
}
\description{
Launches shiny-based interface to configuring and running Citrus. Creates runCitrus.R that can be used to run analysis
in data directory.
}
\examples{
# Uncomment to run
# citrus.launchUI(file.path(system.file(package = "citrus"),"extdata","example1"))
}
\author{
Robert Bruggner
}

