% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.mapToClusterSpace}
\alias{citrus.mapToClusterSpace}
\title{Map new data to existing clusters}
\usage{
citrus.mapToClusterSpace(citrus.combinedFCSSet.new, citrus.combinedFCSSet.old,
  citrus.clustering, mappingColumns = NULL, ...)
}
\arguments{
\item{citrus.combinedFCSSet.new}{A \code{citrus.combinedFCSSet} object containing new data to be mapped.}

\item{citrus.combinedFCSSet.old}{A \code{citrus.combinedFCSSet} object containing new data that has been clustered.}

\item{citrus.clustering}{A clustering of data in \code{citrus.combinedFCSSet.old}.}

\item{mappingColumns}{Parameters to be used for mapping new data to existing cluster space. If \code{NULL}, then parameters
used to cluster \code{citrus.combinedFCSSet.old} are used.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{citrus.mapping} object
\item{clusterMembership}{List of event indices, reporting which events in \code{citrus.combinedFCSSet.new} belong to each cluster in \code{citrus.clustering}.}
\item{mappingColumns}{Columns used to match new data events with existing clustered data.}
}
\description{
Map new data to clusters defined by a clustering.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# List of files to be clustered
fileList1 = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs")[seq(from=2,to=20,by=2)])

# List of files to be mapped
fileList2 = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs")[seq(from=1,to=19,by=2)])

# Read the data
citrus.combinedFCSSet1 = citrus.readFCSSet(dataDirectory,fileList1)
citrus.combinedFCSSet2 = citrus.readFCSSet(dataDirectory,fileList2)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster first dataset
citrus.clustering = citrus.cluster(citrus.combinedFCSSet1,clusteringColumns)

# Map new data to exsting clustering
citrus.mapping = citrus.mapToClusterSpace(citrus.combinedFCSSet.new=citrus.combinedFCSSet1,citrus.combinedFCSSet.old=citrus.combinedFCSSet2,citrus.clustering)
}
\author{
Robert Bruggner
}

