% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Citrus}
\alias{Citrus}
\alias{Citrus-package}
\title{Citrus: Identification of stratifying subpopulations in Flow Cytometry Data}
\description{
Automated identification of clusters whose behavior differs between sample groups.
}

