% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.readFCS}
\alias{citrus.readFCS}
\title{Read an FCS file}
\usage{
citrus.readFCS(filePath, ...)
}
\arguments{
\item{filePath}{The full path to the FCS file to be read.}

\item{...}{Other arguments to be passed to read.FCS.}
}
\value{
a \code{\link{flowCore}} flowFrame object.
}
\description{
Reads an FCS file while suppressing warnings.
}
\author{
Robert Bruggner
}
\seealso{
\code{\link{flowCore}}
}

