% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.readFCSSet}
\alias{citrus.readFCSSet}
\alias{print.citrus.combinedFCSSet}
\title{Read a set of FCS files for analysis}
\usage{
citrus.readFCSSet(dataDirectory, fileList, fileSampleSize = 1000,
  transformColumns = NULL, transformCofactor = 5, scaleColumns = NULL,
  useChannelDescriptions = F, readParameters = NULL, ...)

print.citrus.combinedFCSSet(citrus.combinedFCSSet)
}
\arguments{
\item{dataDirectory}{Full or relative path to directory containin FCS files.}

\item{fileList}{Data frame containing names of files to be read. Each entry should be a file name. FCS files
from the same experimental condition should be listed in the same column with the column name indicating the experimental
condition. FCS files from the same sample measured in multiple experimental conditions should be listed in the same row.
See examples.}

\item{fileSampleSize}{Number of cells to be selected from each FCS file. If this number is larger than the number of events
in a given FCS file, all cells from that file are selected.}

\item{transformColumns}{Vector of parameter names or indicies whose values should transformed prior to analysis.}

\item{scaleColumns}{Vector of parameter names or indicies whose values should be scaled prior to analysis.}

\item{useChannelDescriptions}{Should channel descriptive names be used to name data instead of channel names?}

\item{readParameters}{Vector of parameter names (or descriptions if \code{useChannelDescriptions='T'}) to be read from each FCS file. Can be used to specify common parameters
among FCS files that contain different measured parameters.}

\item{...}{Other undocumented parameters.}

\item{Cofactor}{for arcsin-hyperbolic transform.}
}
\value{
An object of type \code{citrus.combinedFCSSet}.
\item{data}{Combined FCS measurements from files listed in fileList, with paraemters \code{fileEventNumber} and \code{fileId} added. The former
reports the event number in the original source file and the latter records the file from which the event was sampled.}
\item{fileIds}{A matrix recording the assigned fileIds for each file in the fileList argument.}
\item{fileNames}{The corresponding file names for each file, indexed by fileId.}
\item{fileChannelNames}{Names of parameters in each read file.}
\item{fileReagentNames}{Description of parameters in each read file.}
\item{transformColumns}{List of columns that were transformed when read, if supplied.}
\item{transformCofactor}{Cofactor used for \code{arcsinh} transformation, if applicable.}
}
\description{
Reads and combines data from many FCS files for use with Citrus.
}
\examples{
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed - one condition.
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# Create list of files to be analyzed - two conditions.
fileList = data.frame(unstim=list.files(dataDirectory,pattern="unstim"),stim1=list.files(dataDirectory,pattern="stim1"))
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)
}
\author{
Robert Bruggner
}

