% Generated by roxygen2 (4.0.2): do not edit by hand
\name{citrus.selectClusters.minimumClusterSize}
\alias{citrus.selectClusters.minimumClusterSize}
\title{Selects clusters for endpoint analysis}
\usage{
citrus.selectClusters.minimumClusterSize(citrus.clustering,
  minimumClusterSizePercent = 0.05, ...)
}
\arguments{
\item{citrus.clustering}{A \code{citrus.clustering} object.}

\item{minimumClusterSizePercent}{The percentage \eqn{0 < x < 1} of the total number of clustered events a cluster
must contain in order to be selected.}

\item{...}{Other arguments (ignored).}
}
\description{
Selects clusters for endpoint analysis by cluster size. Selected clusters must have a minimum number of
cells in them as a proportion of the total number of clustered events. If \eqn{n} total events are clustered,
clusters contatining at least \eqn{n * minimumClusterSizePercent} events are selected.
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# Cluster data
citrus.clustering = citrus.cluster(citrus.combinedFCSSet,clusteringColumns)

# Select clusters that contain at least 1\% of clustered events.
largeEnoughClusters = citrus.selectClusters.minimumClusterSize(citrus.clustering,minimumClusterSizePercent=0.01)
}
\author{
Robert Bruggner
}
\seealso{
\code{\link{citrus.selectClusters}}
}

