% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{citrus_pair_plot}
\alias{citrus_pair_plot}
\title{Creates pair plot from data table}
\usage{
citrus_pair_plot(model, vars = NULL)
}
\arguments{
\item{model}{list, a citrus segmentation model}

\item{vars}{data.frame, the data to segment}
}
\description{
Creates pair plot from data table
}
