% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_management.R
\name{model_management}
\alias{model_management}
\title{Model management function}
\usage{
model_management(model, hyperparameters)
}
\arguments{
\item{model}{data.frame, the model to save}

\item{hyperparameters}{list, list of hyperparameters of the model}
}
\description{
Saves the model and its settings so that it can be recreated
}
