% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.citrus.full.result}
\alias{plot.citrus.full.result}
\title{Plot a citrus.full.result}
\usage{
plot.citrus.full.result(citrus.full.result, outputDirectory)
}
\arguments{
\item{citrus.full.result}{A citrus.full.result object}

\item{outputDirectory}{Full path to directory in which to place plot output.}
}
\description{
Plot a citrus.full.result
}
\details{
See \code{\link{citrus.full}} for examples.
}
\author{
Robert Bruggner
}

