% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.citrus.regressionResult}
\alias{plot.citrus.regressionResult}
\title{Plot results of a Citrus regression analysis}
\usage{
plot.citrus.regressionResult(citrus.regressionResult, outputDirectory,
  citrus.foldClustering, citrus.foldFeatureSet, citrus.combinedFCSSet,
  plotTypes = c("errorRate", "stratifyingFeatures", "stratifyingClusters",
  "clusterGraph"), hierarchyGraph = NULL, ...)
}
\arguments{
\item{citrus.regressionResult}{A \code{citrus.regressionResult} object.}

\item{outputDirectory}{Full path to output directory for plots.}

\item{citrus.foldClustering}{A \code{citrus.foldClustering} object.}

\item{citrus.foldFeatureSet}{A \code{citrus.foldFeatureSet} object.}

\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object.}

\item{plotTypes}{Vector of plots types to make. Valid options are \code{errorRate} (Cross-validated error rates for predictive models),
\code{stratifyingFeatures} (plots of non-zero model features),\code{stratifyingClusters} (plots of clustering marker distributions in stratifying clusters), and
\code{clusterGraph} (Plots of clustering hierarchy graph).}

\item{hierarchyGraph}{A hierarchy graph configuration created by \code{\link{citrus.createHierarchyGraph}}. If \code{NULL}, automatically generated.}
}
\description{
Makes many plots showing results of a Citrus analysis
}
\examples{
# Where the data lives
dataDirectory = file.path(system.file(package = "citrus"),"extdata","example1")

# Create list of files to be analyzed
fileList = data.frame("unstim"=list.files(dataDirectory,pattern=".fcs"))

# Read the data
citrus.combinedFCSSet = citrus.readFCSSet(dataDirectory,fileList)

# List of columns to be used for clustering
clusteringColumns = c("Red","Blue")

# List disease group of each sample
labels = factor(rep(c("Healthy","Diseased"),each=10))

# Cluster data
citrus.foldClustering = citrus.clusterAndMapFolds(citrus.combinedFCSSet,clusteringColumns,nFolds=1)

# Build abundance features
citrus.foldFeatureSet = citrus.calculateFoldFeatureSet(citrus.foldClustering,citrus.combinedFCSSet)

# Endpoint regress
citrus.regressionResult = citrus.endpointRegress(modelType="pamr",citrus.foldFeatureSet,labels,family="classification")

# Plot results
# plot(citrus.regressionResult,outputDirectory,"/path/to/output/directory/",citrus.foldClustering,citrus.foldFeatureSet,citrus.combinedFCSSet)
}
\author{
Robert Bruggner
}

