% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessed_data.R
\docType{data}
\name{preprocessed_data}
\alias{preprocessed_data}
\title{Segmentation preprocessed data}
\format{Data frame on a customer level. Contains 402 rows and 8 columns.}
\usage{
data(preprocessed_data)
}
\description{
A sample customer dataset for the purpose of demonstrating the segmentation algorithm.
}
\examples{
data(preprocessed_data)
}
\keyword{datasets}
