% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_segment.R
\name{rpart.lists}
\alias{rpart.lists}
\title{rpart.lists function}
\usage{
rpart.lists(object)
}
\arguments{
\item{object}{an rpart object}
}
\value{
a list of lists
}
\description{
THIS HAS BEEN COPIED FROM THE ARCHIVED rpart.utils PACKAGE AND THIS CODE WAS WRITTEN BY THE AUTHORS OF THAT PACKAGE
Creates lists of variable values (factor levels) associated with each rule in an \pkg{rpart} object.
}
\examples{
library(rpart)
fit<-rpart(Reliability~.,data=car.test.frame)
rpart.lists(fit)
}
