% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_segment.R
\name{rpart.rules.table}
\alias{rpart.rules.table}
\title{rpart.rules.table function}
\usage{
rpart.rules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
THIS HAS BEEN COPIED FROM THE ARCHIVED rpart.utils PACKAGE AND THIS CODE WAS WRITTEN BY THE AUTHORS OF THAT PACKAGE
Returns an unpivoted table of branch paths (subrules) associated with each node.
}
\examples{
library(rpart)
fit<-rpart(Reliability~.,data=car.test.frame)
rpart.rules.table(fit)
}
