% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_segment.R
\name{rpart.subrules.table}
\alias{rpart.subrules.table}
\title{rpart.subrules.table function}
\usage{
rpart.subrules.table(object)
}
\arguments{
\item{object}{an rpart object}
}
\description{
THIS HAS BEEN COPIED FROM THE ARCHIVED rpart.utils PACKAGE AND THIS CODE WAS WRITTEN BY THE AUTHORS OF THAT PACKAGE                   
Returns an unpivoted table of variable values (factor levels) associated with each branch.
}
\examples{
library(rpart)
fit<-rpart(Reliability~.,data=car.test.frame)
rpart.subrules.table(fit)
}
