% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactional_data.R
\docType{data}
\name{transactional_data}
\alias{transactional_data}
\title{Segmentation transactional data}
\format{
Data frame on a transactional level. Contains 10,000 rows and 6 columns.
}
\usage{
data(transactional_data)
}
\description{
A sample customer dataset for the purpose of demonstrating the segmentation algorithm.
}
\examples{
data(transactional_data)
}
\keyword{datasets}
