% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_segment.R
\name{tree_abstract}
\alias{tree_abstract}
\title{Abstraction layer function}
\usage{
tree_abstract(model, inputdata)
}
\arguments{
\item{model}{The model to organise}

\item{inputdata}{The data used to train the model}
}
\description{
Organises the model outputs, predictions and settings in a general structure
}
