% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table.R
\name{output_table}
\alias{output_table}
\title{Output Table}
\usage{
output_table(data, model)
}
\arguments{
\item{data}{A dataframe generated from the pre-processing step}

\item{model}{A model object used to classify ids with, generated from the model selection layer}
}
\value{
A tibble providing high-level segment attributes such as mean and max (numeric) or mode (categorical)
for the segmentation features used.
}
\description{
Generates the output table for model and data
}
