% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_segment.R
\name{tree_abstract}
\alias{tree_abstract}
\title{Abstraction layer function}
\usage{
tree_abstract(model)
}
\arguments{
\item{model}{The model to organise}
}
\value{
A structure with the class name "tree_model" which contains a list of all the relevant model data, 
including the rpart model object, hyper-parameters, segment table, labelled customer lookup table, 
and the input data used to train the model.
}
\description{
Organises the model outputs, predictions and settings in a general structure
}
