% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validation function}
\usage{
validate(df, supervised = TRUE, force, hyperparameters)
}
\arguments{
\item{df}{data.frame, the data to validate}

\item{supervised}{logical, TRUE for supervised learning, FALSE for k-clusters}

\item{force}{logical, TRUE to ignore error on categorical columns}

\item{hyperparameters}{list of hyperparameters used in the model}
}
\value{
`TRUE` if all checks are passed. Otherwise an error is raised.
}
\description{
Validates that the input data adheres to the expected format for modelling.
}
